Move server commands to the `client::commands` module
