+++ crates/hyperqueue/src/client/commands/mod.rs
@@ -2,8 +2,8 @@ pub mod autoalloc;
 pub mod event;
 pub mod job;
 pub mod log;
+pub mod server;
 pub mod stats;
-pub mod stop;
 pub mod submit;
 pub mod wait;
 pub mod worker;
