+++ crates/hyperqueue/src/client/mod.rs
@@ -5,6 +5,7 @@ pub mod globalsettings;
 pub mod job;
 pub mod output;
 pub mod resources;
+pub mod server;
 pub mod status;
 pub mod utils;
