+++ crates/hyperqueue/src/client/server.rs
@@ -1,7 +1,7 @@
 use crate::transfer::connection::ClientConnection;
 use crate::transfer::messages::FromClientMessage;
 
-pub async fn stop_server(connection: &mut ClientConnection) -> crate::Result<()> {
+pub async fn client_stop_server(connection: &mut ClientConnection) -> crate::Result<()> {
     connection.send(FromClientMessage::Stop).await?;
     log::info!("Stopping server");
     Ok(())
